separated TodoButtons from TodoItem.jsx into separate file ButtonsTuple.jsx and used icons for buttons content in TodoItem.jsx
