+++ src/TodoItem.jsx
@@ -7,6 +7,10 @@ import {
   View
 } from 'react-native'
 import { colors } from './colors'
+import ButtonsTuple from './ButtonsTuple'
+import XCircleFill from './svg/XCircleFill'
+import PenFill from './svg/PenFill'
+import Save2Fill from './svg/Save2Fill'
 
 function TodoData({ text, date, done }) {
   const dateColorStyle = {
@@ -24,41 +28,6 @@ function TodoData({ text, date, done }) {
           {new Date(date).toLocaleString().slice(0, -3)}
         </Text>
       )}
-    </View>
-  )
-}
-
-function TodoButtons({
-  id,
-  text1,
-  text2,
-  showFirst = true,
-  onPress1,
-  onPress2
-}) {
-  const dynamicBorderRadiusStyle = {
-    borderBottomLeftRadius: showFirst ? 0 : 5,
-    borderTopLeftRadius: showFirst ? 0 : 5
-  }
-
-  return (
-    <View style={style.todoButtons}>
-      {showFirst && (
-        <TouchableOpacity
-          id={id}
-          style={[style.todoButton, style.button1]}
-          onPress={onPress1}
-        >
-          <Text style={style.todoButtonText}>{text1}</Text>
-        </TouchableOpacity>
-      )}
-      <TouchableOpacity
-        id={id}
-        style={[style.todoButton, style.button2, dynamicBorderRadiusStyle]}
-        onPress={onPress2}
-      >
-        <Text style={style.todoButtonText}>{text2}</Text>
-      </TouchableOpacity>
     </View>
   )
 }
@@ -85,10 +54,10 @@ export default function TodoItem({
       {editValue === null ? (
         <>
           <TodoData {...{ text, date, done }} />
-          <TodoButtons
+          <ButtonsTuple
             id={id}
-            text1={'Edit'}
-            text2={'Delete'}
+            button1={<PenFill />}
+            button2={<XCircleFill />}
             showFirst={!done}
             onPress1={onEditButtonPress}
             onPress2={onDeleteButtonPress}
@@ -102,10 +71,10 @@ export default function TodoItem({
             onChangeText={onEditInputChange}
             style={style.editInput}
           />
-          <TodoButtons
+          <ButtonsTuple
             id={id}
-            text1={'Save'}
-            text2={'Cancel'}
+            button1={<Save2Fill />}
+            button2={<XCircleFill />}
             onPress1={onSaveButtonPress}
             onPress2={onCancelButtonPress}
           />
@@ -143,27 +112,5 @@ const style = StyleSheet.create({
   },
   date: {
     fontSize: 11
-  },
-  todoButtons: {
-    flexDirection: 'row',
-    marginLeft: 10
-  },
-  todoButton: {
-    padding: 5,
-    paddingLeft: 7,
-    paddingRight: 7,
-    borderRadius: 5
-  },
-  todoButtonText: {
-    color: 'white',
-    fontSize: 18
-  },
-  button1: {
-    backgroundColor: colors.buttonPink,
-    borderBottomRightRadius: 0,
-    borderTopRightRadius: 0
-  },
-  button2: {
-    backgroundColor: colors.buttonPurple
   }
 })
